pyproject.toml
[tool.poetry]
name = "gradio-langchain"
version = "0.1.0"
description = ""
authors = ["MilaisTaso <livreblanche819@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.21.0"
langchain = "^0.1.12"
python-dotenv = "^1.0.1"
openai = "^1.14.0"
langchain-openai = "^0.0.8"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
